New folder created to define ORM classes.
The SQLAlchemy ORM classes are mapped to the tables in the SQLite database.

These ORM classes may be used to interact with the database tables using SQLAlchemy ORM.

Even if SQLAlchemy Core is used (instead of ORM objects), these ORM classes provide documentation of the database schema.

Each file contains an ORM class that maps to a single database table.

"columns()" methods are added to each class for easily obtaining a list of table columns.

"column_map()" methods are added to generate a dictionary where keys represent the column names of DataFrames containing web-scraped data,
and values represent the corresponding column names in the database table. This ensures that data is inserted into the correct columns of the SQLite tables.

Dictionaries generated by column_map() are wrapped with MappingProxyType to prevent accidental mutation.