Changed _split_units method
Previously, regex was used to separate units from commodity names, because the read_html method of pandas
joins the commodity names and units with a whitespace. This regex approach failed in some cases.

Now, BeautifulSoup is used to extract units from HTML (separated from commodity names via \n\n so easier to extract).
Then, the dataframe is revised to give a column with just commodity names, and a column with precise units.